extracting UI element views to separate file
